Se arreglo el problema que habia al querer vender sin estar logueado , ahora si no esta logueado e interactua con el boton de vender le redirijira a la pantalla del login y si esta logueado pues le aparecera la ventana de vender
